+++ test/unit/common_test.rb
@@ -56,27 +56,6 @@ class AppiumLibCoreTest
       }.freeze
 
       def test_create_session_force_mjsonwp
-        response = {
-          status: 0, # To make bridge.dialect == :oss
-          value: RESPONSE_BASE_VALUE
-        }.to_json
-
-        stub_request(:post, 'http://127.0.0.1:4723/wd/hub/session')
-          .with(body: { desiredCapabilities: CAPS }.to_json)
-          .to_return(headers: Mock::HEADER, status: 200, body: response)
-
-        stub_request(:post, "#{Mock::SESSION}/timeouts/implicit_wait")
-          .with(body: { ms: 0 }.to_json)
-          .to_return(headers: Mock::HEADER, status: 200, body: { value: nil }.to_json)
-
-        driver = ::Appium::Core.for({ caps: CAPS.merge({ forceMjsonwp: true }), appium_lib: {} }).start_driver
-
-        assert_requested(:post, 'http://127.0.0.1:4723/wd/hub/session', times: 1)
-        assert_requested(:post, "#{Mock::SESSION}/timeouts/implicit_wait", body: { ms: 0 }.to_json, times: 1)
-        driver
-      end
-
-      def test_create_session_force_mjsonwp_false
         response = { value: RESPONSE_BASE_VALUE }.to_json
 
         stub_request(:post, 'http://127.0.0.1:4723/wd/hub/session')
@@ -88,52 +67,11 @@ class AppiumLibCoreTest
           .with(body: { implicit: 0 }.to_json)
           .to_return(headers: Mock::HEADER, status: 200, body: { value: nil }.to_json)
 
-        driver = ::Appium::Core.for({ caps: CAPS.merge({ forceMjsonwp: false }), appium_lib: {} }).start_driver
+        driver = ::Appium::Core.for({ caps: CAPS.merge({ forceMjsonwp: true }), appium_lib: {} }).start_driver
 
         assert_requested(:post, 'http://127.0.0.1:4723/wd/hub/session', times: 1)
         assert_requested(:post, "#{Mock::SESSION}/timeouts", body: { implicit: 0 }.to_json, times: 1)
         driver
-      end
-
-      def test_create_session_force_mjsonwp_with_source_package
-        response = {
-          status: 0, # To make bridge.dialect == :oss
-          value: {
-            sessionId: '1234567890',
-            capabilities: {
-              platformName: :android,
-              automationName: 'uiautomator2',
-              app: 'sauce-storage:test/functional/app/api.apk.zip',
-              platformVersion: '7.1.1',
-              deviceName: 'Android Emulator',
-              appPackage: 'io.appium.android.apis'
-            }
-          }
-        }.to_json
-        http_caps = {
-          platformName: :android,
-          automationName: 'uiautomator2',
-          app: 'sauce-storage:test/functional/app/api.apk.zip',
-          platformVersion: '7.1.1',
-          deviceName: 'Android Emulator',
-          appPackage: 'io.appium.android.apis'
-        }
-
-        stub_request(:post, 'http://127.0.0.1:4723/wd/hub/session')
-          .with(body: { desiredCapabilities: http_caps }.to_json)
-          .to_return(headers: Mock::HEADER, status: 200, body: response)
-
-        stub_request(:post, "#{Mock::SESSION}/timeouts/implicit_wait")
-          .with(body: { ms: 0 }.to_json)
-          .to_return(headers: Mock::HEADER, status: 200, body: { value: nil }.to_json)
-
-        core = ::Appium::Core.for({ caps: http_caps.merge({ forceMjsonwp: true }), appium_lib: {} })
-        core.start_driver
-
-        assert_requested(:post, 'http://127.0.0.1:4723/wd/hub/session', times: 1)
-        assert_requested(:post, "#{Mock::SESSION}/timeouts/implicit_wait", body: { ms: 0 }.to_json, times: 1)
-
-        assert_equal 'sauce-storage:test/functional/app/api.apk.zip', core.caps[:app]
       end
 
       def test_create_session_w3c